Change data type

+++ app/src/main/java/gachon/teama/frimo/data/remote/Diary.kt
@@ -6,11 +6,10 @@ import androidx.room.PrimaryKey
 import kotlinx.parcelize.Parcelize
 
 // Todo: 키워드, 댓글 추가 방법 알아볼 것
-// Todo: diary id Int type에서 Long type으로 변경
 @Entity
 @Parcelize
 data class Diary(
-    @PrimaryKey val id: Int, // diary 구분자
+    @PrimaryKey val id: Long, // diary 구분자
     val title: String, // diary 재목
     val content: String, // diary 내용
     val created: String, // diary 작성 날짜
